remove unused method, reorg unit tests appropriately
